feat: incorporate `huggingface` spaces
