+++ .dockerignore
@@ -1,0 +1,7 @@
+# Ignore everything
+**
+
+# ... except for dependencies and source
+!pyproject.toml
+!poetry.lock
+!src
