--- .github/workflows/auto-approve.yml
@@ -1,29 +1,0 @@
-name: Auto approve
-on: pull_request_target
-jobs:
-  auto-approve:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: hmarr/auto-approve-action@v2
-        if: |
-          (
-            github.event.pull_request.user.login == 'dependabot[bot]' ||
-            github.event.pull_request.user.login == 'dependabot' ||
-            github.event.pull_request.user.login == 'dependabot-preview[bot]' ||
-            github.event.pull_request.user.login == 'dependabot-preview' ||
-            github.event.pull_request.user.login == 'renovate[bot]' ||
-            github.event.pull_request.user.login == 'renovate' ||
-            github.event.pull_request.user.login == 'github-actions[bot]'
-          )
-            &&
-          (
-            github.actor == 'dependabot[bot]' ||
-            github.actor == 'dependabot' ||
-            github.actor == 'dependabot-preview[bot]' ||
-            github.actor == 'dependabot-preview' ||
-            github.actor == 'renovate[bot]' ||
-            github.actor == 'renovate' ||
-            github.actor == 'github-actions[bot]'
-          )
-        with:
-          github-token: ${{ secrets.GITHUB_TOKEN }}
