+++ pyproject.toml
@@ -6,7 +6,7 @@ requires = [
 
 [tool.black]
 line-length = 99
-target-version = ["py37", "py38"]
+target-version = ["py39"]
 skip-string-normalization = true
 
 [tool.isort]
@@ -88,9 +88,10 @@ ray = ">=2.5.1"
 click = ">=8.1.3"
 rich = ">=13.4.2"
 pyinsect = {git = "https://github.com/ggianna/PyINSECT"}
+kaleido = "0.2.1"
 torch = "1.9.0"
-torchvision = "0.15.2"
-kaleido = "0.2.1.post1"
+torchvision = "0.10.0"
+gradio = "^3.39.0"
 
 [tool.poetry.group.dev.dependencies]
 python-semantic-release = "*"
@@ -155,7 +156,7 @@ paths = ["src/roughgan", "tests"]
     help = "Build and serve the documentation"
 
     [[tool.poe.tasks.docs.sequence]]
-      ref = "_open --url http://127.0.0.1:8000/Alpaca-with-a-Mixture-of-Teachers/"
+      ref = "_open --url http://127.0.0.1:8000/RoughGAN/"
 
     [[tool.poe.tasks.docs.sequence]]
       cmd = "poetry run mkdocs serve --livereload --watch-theme"
@@ -164,21 +165,13 @@ paths = ["src/roughgan", "tests"]
     help = "Export dependencies to requirements.txt format"
     cmd = "poetry export --without-hashes --dev -o requirements.txt"
 
-  [tool.poe.tasks.docker-build-api]
-    help = "Build the API docker image"
-    shell = "docker build . -t roughgan:$( git tag -l | tail -1 | cut -c2- ) -t build:api -f ./docker/Dockerfile"
-
-  [tool.poe.tasks.docker-build-train]
+  [tool.poe.tasks.docker-build]
     help = "Build the model training docker image"
-    shell = "docker build . -t roughgan:$( git tag -l | tail -1 | cut -c2- ) -t build:train -f ./docker/Dockerfile.train"
+    shell = "docker build . -t roughgan:$( git tag -l | tail -1 | cut -c2- ) -t build:train -f Dockerfile"
 
-  [tool.poe.tasks.docker-run-train]
+  [tool.poe.tasks.docker-run]
     help = "Train the model using Docker"
-    shell = "docker run -v $(pwd)/data:/home/app/app/data -v $(pwd)/model:/home/app/app/model --gpus $(nvidia-smi --list-gpus | wc -l) roughgan:$( git tag -l | tail -1 | cut -c2- )"
-
-  [tool.poe.tasks.docker-run-api]
-    help = "Run the docker image as a container"
-    shell = "docker run -p 8080:8080 mlogo:$( git tag -l | tail -1 | cut -c2- )"
+    shell = "docker run -v $(pwd)/data:/home/app/app/data -v $(pwd)/models:/home/app/app/model --gpus $(nvidia-smi --list-gpus | wc -l) roughgan:$( git tag -l | tail -1 | cut -c2- )"
 
   [tool.poe.tasks._open]
     help = "Open a URL"
@@ -225,7 +218,7 @@ line-length = 99
 
 dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
 
-target-version = "py310"
+target-version = "py39"
 
 [tool.ruff.mccabe]
 max-complexity = 10
